adding in main changes to axum 0.6 (#29)
* remove double quote from column in mysql (#25)

Mysql does not support double quoting column names like postgresql does.

* add Debug and Clone derive to AxumNullPool (#27)

In the examples, it is shown that we can use AxumNullPool to avoid using DB (useful for unit tests). But based on the docs/source code AxumSessionStore requires that the client `T: AxumDatabasePool + Clone + Debug + Sync + Send + 'static`
The issue is that AxumNullPool is neither Clone nor Debug. So example with AxumNullPool actually does not work, I also tried it in on my own. There is no reason for AxumNullPool not to be Clone and or Debug

This is simple fix by adding derive with Clone  and Debug

Co-authored-by: tadic-luka <luka.tadic365@gmail.com>